changed price in both parts and wishlist parts tables to a float instead of an integer.

+++ van-builder/db/migrate/20190313001019_add_wishlist_parts_table.rb
@@ -2,7 +2,7 @@ class AddWishlistPartsTable < ActiveRecord::Migration
   def change
     create_table :wishlist_parts do |t|
       t.string  :name
-      t.integer :price
+      t.float :price
       t.string  :description
       t.string  :manufacturer
       t.integer :user_id
